Working on IFC properties management
