+++ src/core/index.ts
@@ -10,7 +10,6 @@ export * from "./SimpleClipper";
 export * from "./ScreenCuller";
 export * from "./SimpleDimensions";
 export * from "./LocalCacher";
-export * from "./PropertiesProcessor";
 export * from "./SimpleAngle";
 export * from "./SimpleArea";
 export * from "./SimpleSVGViewport";
