+++ src/fragments/FragmentTree/index.ts
@@ -1,29 +1,33 @@
 import { Component, UI } from "../../base-types";
 import { TreeView } from "../../ui";
-import { FragmentTreeItem } from "./tree-item";
+import { FragmentTreeItem } from "./src/tree-item";
 import { Components } from "../../core";
-import { FragmentGrouper } from "../fragment-grouper";
+import { FragmentGrouper } from "../FragmentGrouper";
 import { FragmentHighlighter } from "../FragmentHighlighter";
 
 export class FragmentTree extends Component<FragmentTreeItem> implements UI {
   uiElement: TreeView;
   name: string;
   enabled: boolean = true;
-  private _components: Components;
   groupSystemNames: string[];
   functionsMap: { [groupSystemName: string]: () => void } = {};
+  private _components: Components;
+  private _fragmentHighlighter: FragmentHighlighter;
+  private _fragmentGrouper: FragmentGrouper;
 
   private readonly _tree: FragmentTreeItem;
 
   constructor(
     components: Components,
-    private _fragmentHighlighter: FragmentHighlighter,
-    private _fragmentGrouper: FragmentGrouper,
+    fragmentHighlighter: FragmentHighlighter,
+    fragmentGrouper: FragmentGrouper,
     name: string,
     groupSystemNames: string[]
   ) {
     super();
     this._components = components;
+    this._fragmentGrouper = fragmentGrouper;
+    this._fragmentHighlighter = fragmentHighlighter;
     this.name = name;
     this.groupSystemNames = groupSystemNames;
 
@@ -64,7 +68,7 @@ export class FragmentTree extends Component<FragmentTreeItem> implements UI {
       // name is N00, N01, N02...
       const filter = { ...result, [currentSystemName]: name }; // { storeys: "N00" }, { storeys: "N01" }...
       const hasElements =
-        Object.keys(this._fragmentGrouper.get(filter)).length > 0;
+        Object.keys(this._fragmentGrouper.getByFilter(filter)).length > 0;
       if (hasElements) {
         const treeItemName =
           currentSystemName[0].toUpperCase() + currentSystemName.slice(1); // Storeys
