+++ src/fragments/index.ts
@@ -2,7 +2,8 @@ export * from "./FragmentManager";
 export * from "./FragmentIfcLoader";
 export * from "./FragmentHighlighter";
 export * from "./FragmentTree";
-export * from "./fragment-grouper";
+export * from "./FragmentGrouper";
 export * from "./fragment-edges";
 export * from "./FragmentGroups";
 export * from "./FragmentCacher";
+export * from "./FragmentCoordinator";
