+++ src/ifc/PropertiesProcessor/index.html
@@ -1,0 +1,138 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport"
+        content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <link rel="stylesheet" href="../../../resources/styles.css">
+  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+  <link rel="stylesheet" href="../../ui/UIManager/top.css">
+  <link rel="icon" type="image/x-icon" href="../../../resources/favicon.ico">
+  <title>PropertiesProcessor Component</title>
+</head>
+<body>
+<div class="full-screen" id="container">
+  <div id="properties-container" class="simple-container"></div>
+</div>
+<script type="importmap">
+    {
+      "imports": {
+        "three": "https://unpkg.com/three@0.152.2/build/three.module.js",
+        "stats.js/src/Stats.js": "https://unpkg.com/stats-js@1.0.1/src/Stats.js",
+        "three/examples/jsm/libs/lil-gui.module.min": "https://unpkg.com/three@0.152.2/examples/jsm/libs/lil-gui.module.min.js",
+        "openbim-components": "../../../resources/openbim-components.js"
+      }
+    }
+
+</script>
+<script type="module">
+
+	// Set up scene (see SimpleScene tutorial)
+
+	import * as THREE from 'three';
+	import * as OBC from 'openbim-components';
+	import Stats from 'stats.js/src/Stats.js';
+	import * as dat from 'three/examples/jsm/libs/lil-gui.module.min';
+
+	const container = document.getElementById('container');
+
+	const components = new OBC.Components();
+
+	components.scene = new OBC.SimpleScene(components);
+	components.renderer = new OBC.SimpleRenderer(components, container);
+	components.camera = new OBC.SimpleCamera(components);
+	components.raycaster = new OBC.SimpleRaycaster(components);
+
+	components.init();
+
+	const scene = components.scene.get();
+
+	scene.background = null;
+	components.camera.controls.setLookAt(10, 10, 10, 0, 0, 0);
+
+	const directionalLight = new THREE.DirectionalLight();
+	directionalLight.position.set(5, 10, 3);
+	directionalLight.intensity = 0.5;
+	scene.add(directionalLight);
+
+	const ambientLight = new THREE.AmbientLight();
+	ambientLight.intensity = 0.5;
+	scene.add(ambientLight);
+
+	const grid = new OBC.SimpleGrid(components);
+	components.tools.add(grid);
+
+	const fragments = new OBC.FragmentManager(components);
+	const fragmentIfcLoader = new OBC.FragmentIfcLoader(components, fragments);
+
+	//#region Setup elements highlight
+	const highlighter = new OBC.FragmentHighlighter(components, fragments);
+	highlighter.enabled = true;
+	highlighter.add('select', [new THREE.MeshLambertMaterial({ color: 'gold' })]);
+	highlighter.add('highlight', [new THREE.MeshLambertMaterial({ color: 'gold', transparent: true, opacity: 0.3 })]);
+	container.addEventListener('click', e => {
+		if (e.target !== components.renderer.get().domElement) {
+			return;
+		}
+		highlighter.highlight('select');
+	});
+	container.addEventListener('mousemove', e => {
+		if (e.target !== components.renderer.get().domElement) {
+			return;
+		}
+		const highlight = highlighter.highlight('highlight');
+	});
+	//#endregion
+
+	const propertiesProcessor = new OBC.PropertiesProcessor(components, fragments, highlighter);
+	const propertiesContainer = document.getElementById('properties-container');
+	propertiesContainer.append(propertiesProcessor.uiElement.get());
+
+	fragmentIfcLoader.settings.wasm = {
+		path: '../../../resources/',
+		absolute: false,
+	};
+
+	fragmentIfcLoader.settings.webIfc.COORDINATE_TO_ORIGIN = true;
+	fragmentIfcLoader.settings.webIfc.USE_FAST_BOOLS = false;
+
+	async function loadIfcAsFragments(url) {
+		const file = await fetch(url);
+		const data = await file.arrayBuffer();
+		const buffer = new Uint8Array(data);
+		const model = await fragmentIfcLoader.load(buffer);
+
+		propertiesProcessor.process(model);
+
+		scene.add(model);
+		for (const fragment of model.fragments) {
+			fragments.list[fragment.id] = fragment;
+		}
+
+		highlighter.update();
+		grid.visible = false;
+
+	}
+
+	// Set up stats
+
+	const stats = new Stats();
+	stats.showPanel(2);
+	document.body.append(stats.dom);
+	stats.dom.style.left = '0px';
+	const renderer = components.renderer;
+	renderer.beforeUpdate.on(() => stats.begin());
+	renderer.afterUpdate.on(() => stats.end());
+
+	// Set up action buttons
+
+	const toolbar = new OBC.Toolbar(components);
+	const button = new OBC.Button(components, { name: 'Load IFC', materialIconName: 'upload_file' });
+	button.onclick = () => loadIfcAsFragments('../../../resources/asdf2.ifc');
+	toolbar.addButton(button);
+	components.ui.addToolbar(toolbar);
+
+</script>
+</body>
+</html>
