+++ src/ifc/PropertiesProcessor/src/element-spatial-container.ts
@@ -1,0 +1,10 @@
+import * as WEBIFC from "web-ifc";
+
+export function getElementSpatialContainer(
+  properties: Record<string, Record<string, any>>
+) {
+  const arrayProperties = Object.values(properties);
+  arrayProperties.filter((prop: any) => {
+    return prop.type === WEBIFC.IFCRELCONTAINEDINSPATIALSTRUCTURE;
+  });
+}
