+++ src/ifc/PropertiesProcessor/src/index.ts
@@ -1,0 +1,5 @@
+export * from "./element-psets";
+export * from "./element-qsets";
+export * from "./pset-props";
+export * from "./qset-quantities";
+export * from "./relation-map";
