+++ src/ifc/index.ts
@@ -3,3 +3,5 @@ export * from "./ifc-category-map";
 export * from "./ifc-elements-map";
 export * from "./ifc-json-exporter";
 export * from "./ifc-geometry-types";
+export * from "./IfcPropertiesManager";
+export * from "./PropertiesProcessor";
